requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==2.3.7
qrcode==7.4.2
Pillow>=10.0.0
python-dateutil==2.8.2
